requirements.txt
streamlit
earthengine-api
numpy
pandas
matplotlib
scikit-learn
google-auth